requirements.txt

Flask==2.3.2
pytesseract==0.3.10
Pillow==10.0.0
gunicorn==20.1.0
requests==2.32.3
flask-cors==4.0.0
python-bidi==0.6.6
arabic-reshaper==3.0.0